requirements.txt
streamlit==1.36.0
streamlit-extras==0.4.3
civitai-py==0.1.10
httpx==0.27.0